Split repository base into two classes
