+++ backend/infrahub/git/__init__.py
@@ -1,23 +1,11 @@
 from infrahub.git.directory import initialize_repositories_directory
 from infrahub.git.repository import (
-    ArtifactGenerateResult,
-    CheckDefinitionInformation,
-    GraphQLQueryInformation,
     InfrahubReadOnlyRepository,
     InfrahubRepository,
-    RepoFileInformation,
-    TransformPythonInformation,
-    extract_repo_file_information,
 )
 
 __all__ = [
-    "ArtifactGenerateResult",
     "InfrahubReadOnlyRepository",
     "InfrahubRepository",
-    "TransformPythonInformation",
-    "CheckDefinitionInformation",
-    "RepoFileInformation",
-    "GraphQLQueryInformation",
-    "extract_repo_file_information",
     "initialize_repositories_directory",
 ]
