+++ pyproject.toml
@@ -273,6 +273,14 @@ module = "infrahub.database"
 ignore_errors = true
 
 [[tool.mypy.overrides]]
+module = "infrahub.git.base"
+ignore_errors = true
+
+[[tool.mypy.overrides]]
+module = "infrahub.git.integrator"
+ignore_errors = true
+
+[[tool.mypy.overrides]]
 module = "infrahub.git.repository"
 ignore_errors = true
